this commit will add error handling in boardRoutes and cardRoutes

+++ backend/routes/boardRoutes.js
@@ -16,22 +16,38 @@ router.use('/', cardRoutes)
 // get all boards from the database using routes
 // since were interacting with a database, use async/await
 router.get('/', async (req, res) => {
-    const boards = await prisma.Board.findMany({include: {cards: true}});
-    // shows the relation (cards array)
-    res.json(boards);
+    try {
+        const boards = await prisma.Board.findMany(
+            // {include: {cards: true}}
+        );
+        // shows the relation (cards array)
+        res.json(boards);
+    } catch (error) {
+        res.status(500).send('Server Error');
+    }
 })
 
 // get board by id
 router.get('/:boardId', async (req, res) => {
     const boardId = parseInt(req.params.boardId)
-    const board = await prisma.Board.findUnique({
-        where: {id: boardId},
-        include: {cards: true}
-    })
-    res.json(board);
+    try {
+        const board = await prisma.Board.findUnique({
+            where: {id: boardId},
+            include: {cards: true}
+        })
+
+        if (board) {
+            res.json(board);
+        } else {
+            res.status(404).send("Board not found")
+        }
+        
+    } catch (error) {
+        res.status(500).send('An error occurred while fetching the board')
+    }
 })
 
-// post request, listens for create requests
+// create a new board
 router.post('/', async (req, res) => {
     if (!req.body.title || !req.body.description || !req.body.category || !req.body.image) {
         return res.status(400).send("title, description, category, and image are required")
@@ -39,16 +55,20 @@ router.post('/', async (req, res) => {
     const {title, description, category, image, author, cards} = req.body;
 
     // create a new board object
-    const newBoard = await prisma.board.create({
-        data: {
-            title,
-            description,
-            category,
-            image,
-            author,
-        }
-    })
-    res.json(newBoard)
+    try {
+        const newBoard = await prisma.board.create({
+            data: {
+                title,
+                description,
+                category,
+                image,
+                author,
+            }
+        })
+        res.json(newBoard)
+    } catch (error) {
+        res.status(500).send('An error occurred while creating the board')
+    }
 })
 
 // update a board element
@@ -57,26 +77,50 @@ router.put('/:boardId', async (req, res) => {
         return res.status(400).send("title, description, category, and image are required")
     }
     const boardId = parseInt(req.params.boardId)
-    const {title, description, category, image, author, cards} = req.body;
-    const updatedBoard = await prisma.board.update({
-        where: {id: boardId},
-        data: {
-            title,
-            description,
-            category,
-            image,
-            author
+    const {title, description, category, image, author} = req.body;
+    try {    
+        const checkBoardExists = await prisma.Board.findUnique({
+            where: {id: boardId},
+        })
+
+        if (!checkBoardExists) {
+            return res.status(404).send("Board not found")
         }
-    })
-    res.json(updatedBoard)
+
+        const updatedBoard = await prisma.board.update({
+            where: {id: boardId},
+            data: {
+                title,
+                description,
+                category,
+                image,
+                author
+            }
+        })
+        res.json(updatedBoard)
+    } catch (error) {
+        res.status(500).send("An error occurred while updating the board")
+    }
 })
 
 router.delete('/:boardId', async (req, res) => {
     const boardId = parseInt(req.params.boardId)
-    const deletedBoard = await prisma.board.delete({
-        where: {id: boardId}
-    })
-    res.json(deletedBoard);
+    try {
+        const checkBoardExists = await prisma.Board.findUnique({
+            where: {id: boardId},
+        })
+
+        if (!checkBoardExists) {
+            return res.status(404).send("Board not found")
+        }
+        
+        const deletedBoard = await prisma.board.delete({
+            where: {id: boardId}
+        })
+        res.json(deletedBoard);
+    } catch (error) {
+        res.status(500).send("An error occurred while deleting the board")
+    }
 })
 
 module.exports = router;
